requirements.txt
numpy
pandas
faiss-cpu
scikit-learn
sentence-transformer